#22 implement logging interfaces for microsoft
